refactor: split models.go into constants.go and types.go
